Refactor Mask in Attention
Currently, the attention code is **hardcoded** to handle either `causal_mask`
or an arbitrary `mask_fn`.

To support **sliding window masks**, we previously used a **hack** by injecting
the `_sliding_window_size` attribute into functions.

This refactor **makes the masking logic more flexible** by allowing arbitrary
`MaskFnAttentionBias`.
- If downstream requires a **new mask pattern**, they can simply:
  1. Implement a **subclass of `MaskFnAttentionBias`**.
  2. Set `attention.mask` accordingly.

+++ axlearn/common/lora_test.py
@@ -263,7 +263,7 @@ class LoraFusedQKVLinearTest(TestCase):
 
         initial_cache_state, init_output = layer.init_states(
             time_step=None,
-            query=TensorSpec([batch_size, seq_len]),
+            query=TensorSpec([batch_size, seq_len], dtype=q_proj.dtype),
         )
         self.assertIsNone(init_output)
